broke avion into separate files, optimization in the engine
